bugfix: don't check quorum for evaluate stage proposals (#177)
* bugfix: don't check quorum for evaluate stage proposals

the quorum is dynamic and always changing depending on the number of proposals, number of eligible voters, and their vote power. Therefore we only check quorum on the initial vote. Once a proposal has been voted in and is in eval stage, we only check unity, in case someone changed their mind. But we don't check quorum.

* checkeval action added

* removed checkeval action again